feat: 커스텀 훅간의 의존성 제거 및 SearchPage 렌더링 방식 변경
