+++ apps/client/src/pages/search/hooks/use-search-logic.ts
@@ -1,36 +1,22 @@
 import { useState } from 'react';
-import { useNavigate, useSearchParams } from 'react-router-dom';
-
-import { routePath } from '@shared/constants/path';
-
-import { useSearchArtist } from '../hooks/use-search-data';
+import { useNavigate } from 'react-router-dom';
 
 export const useSearchLogic = () => {
   const [barFocus, setBarFocus] = useState(false);
   const navigate = useNavigate();
-  const [searchParams] = useSearchParams();
-  const paramsKeyword = searchParams.get('q') || '';
-
-  const { data: searchData, isLoading } = useSearchArtist({
-    keyword: paramsKeyword,
-    enabled: !!paramsKeyword,
-  });
 
   const handleOnFocus = () => setBarFocus(true);
   const handleOnBlur = () => setBarFocus(false);
 
   const navigateWithKeyword = (keyword: string) => {
-    navigate(`${routePath.SEARCH}?q=${keyword}`);
+    navigate(`/search?q=${keyword}`);
     setBarFocus(false);
   };
 
   return {
-    artistData: searchData?.artist,
-    paramsKeyword,
     barFocus,
     handleOnFocus,
     handleOnBlur,
     navigateWithKeyword,
-    isLoading,
   };
 };
